+++ src/nts_processing/destination_choice_ml_modelling/main_run.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+"""
+Created on: 9/11/2024
+Original author: Adil Zaheer
+"""
+# pylint: disable=import-error,wrong-import-position
+# pylint: enable=import-error,wrong-import-position
+
+
+def main(params):
+
+
+
+
+
+    return
